refactoring of ssr example
