+++ examples/server-side-rendering/backend/api/index.ts
@@ -1,4 +1,5 @@
 import { type Express, type Request, type Response, type NextFunction } from 'express';
+
 import { type Api, type ProductBrief, type Product } from '../../contract';
 
 // sumulates the database
